refactor: tweak examples
